+++ test/audiovideocontroller/DefaultAudioVideoController.test.ts
@@ -4,6 +4,7 @@
 import * as chai from 'chai';
 import * as sinon from 'sinon';
 
+import Attendee from '../../src/attendee/Attendee';
 import DefaultAudioVideoController from '../../src/audiovideocontroller/DefaultAudioVideoController';
 import AudioVideoObserver from '../../src/audiovideoobserver/AudioVideoObserver';
 import Backoff from '../../src/backoff/Backoff';
@@ -26,10 +27,14 @@ import {
   SdkJoinAckFrame,
   SdkLeaveAckFrame,
   SdkSignalFrame,
+  SdkStreamDescriptor,
+  SdkStreamMediaType,
   SdkSubscribeAckFrame,
   SdkTurnCredentials,
 } from '../../src/signalingprotocol/SignalingProtocol.js';
+import SimulcastLayers from '../../src/simulcastlayers/SimulcastLayers';
 import AllHighestVideoBandwidthPolicy from '../../src/videodownlinkbandwidthpolicy/AllHighestVideoBandwidthPolicy';
+import VideoSource from '../../src/videosource/VideoSource';
 import NScaleVideoUplinkBandwidthPolicy from '../../src/videouplinkbandwidthpolicy/NScaleVideoUplinkBandwidthPolicy';
 import DefaultWebSocketAdapter from '../../src/websocketadapter/DefaultWebSocketAdapter';
 import DOMMockBehavior from '../dommock/DOMMockBehavior';
@@ -102,8 +107,11 @@ describe('DefaultAudioVideoController', () => {
     return joinAckSignalBuffer;
   }
 
-  function makeIndexFrame(): Uint8Array {
+  function makeIndexFrame(sources?: SdkStreamDescriptor[]): Uint8Array {
     const indexFrame = SdkIndexFrame.create();
+    if (sources) {
+      indexFrame.sources = sources;
+    }
     const indexSignal = SdkSignalFrame.create();
     indexSignal.type = SdkSignalFrame.Type.INDEX;
     indexSignal.index = indexFrame;
@@ -113,6 +121,17 @@ describe('DefaultAudioVideoController', () => {
     indexSignalBuffer.set(buffer, 1);
     return indexSignalBuffer;
   }
+
+  const makeIndexFrameWithAttendees = (attendees: Attendee[]): Uint8Array => {
+    const sources = attendees.map(({ attendeeId, externalUserId }) => {
+      const streamDescriptor = SdkStreamDescriptor.create();
+      streamDescriptor.attendeeId = attendeeId;
+      streamDescriptor.externalUserId = externalUserId;
+      streamDescriptor.mediaType = SdkStreamMediaType.VIDEO;
+      return streamDescriptor;
+    });
+    return makeIndexFrame(sources);
+  };
 
   // For SubscribeAndReceiveSubscribeAckTask
   function makeSubscribeAckFrame(): Uint8Array {
@@ -1744,6 +1763,170 @@ describe('DefaultAudioVideoController', () => {
       await sendICEEventAndSubscribeAckFrame();
 
       await stop();
+    });
+  });
+
+  describe('getRemoteVideoSources', () => {
+    const compare = (a: VideoSource, b: VideoSource): number =>
+      a.attendee.attendeeId.localeCompare(b.attendee.attendeeId);
+
+    it('should match index frame sources excluding self', async () => {
+      const expectedVideoSources = [{ attendee: { attendeeId: 'a', externalUserId: 'a' } }];
+      const attendeesToMakeIndexFrame = [...expectedVideoSources].map(
+        VideoSource => VideoSource.attendee
+      );
+      audioVideoController = new DefaultAudioVideoController(
+        configuration,
+        new NoOpDebugLogger(),
+        webSocketAdapter,
+        new NoOpDeviceController(),
+        reconnectController
+      );
+      await start();
+      await delay(300);
+      webSocketAdapter.send(makeIndexFrameWithAttendees(attendeesToMakeIndexFrame));
+      await delay(300);
+      const receivedVideoSources = audioVideoController.getRemoteVideoSources();
+      expect(receivedVideoSources.sort(compare)).to.eql(expectedVideoSources.sort(compare));
+      await stop();
+    });
+
+    it('should return an array of length 0, when videoStreamIndex is not initialized', async () => {
+      audioVideoController = new DefaultAudioVideoController(
+        configuration,
+        new NoOpDebugLogger(),
+        webSocketAdapter,
+        new NoOpDeviceController(),
+        reconnectController
+      );
+      const videoSources = audioVideoController.getRemoteVideoSources();
+      expect(videoSources).to.have.lengthOf(0);
+    });
+
+    it('should return [] when called after meeting session is stopped', async () => {
+      audioVideoController = new DefaultAudioVideoController(
+        configuration,
+        new NoOpDebugLogger(),
+        webSocketAdapter,
+        new NoOpMediaStreamBroker(),
+        reconnectController
+      );
+      await start();
+      await stop();
+      await delay(500);
+      expect(audioVideoController.getRemoteVideoSources()).to.deep.equal([]);
+    });
+  });
+
+  describe('encodingSimulcastLayersDidChange', () => {
+    it('observer should get called only when added', async () => {
+      configuration.enableUnifiedPlanForChromiumBasedBrowsers = true;
+      configuration.enableSimulcastForUnifiedPlanChromiumBasedBrowsers = true;
+      domMockBehavior.browserName = 'chrome';
+      domMockBuilder = new DOMMockBuilder(domMockBehavior);
+      audioVideoController = new DefaultAudioVideoController(
+        configuration,
+        new NoOpDebugLogger(),
+        webSocketAdapter,
+        new NoOpDeviceController(),
+        reconnectController
+      );
+      // @ts-ignore
+      expect(audioVideoController.enableSimulcast).to.equal(true);
+      let event = 0;
+      let observed = 0;
+      class TestObserver implements AudioVideoObserver {
+        encodingSimulcastLayersDidChange(_simulcastLayers: SimulcastLayers): void {
+          if (event !== 1 && event !== 3 && event !== 5) {
+            assert.fail();
+          }
+          observed += 1;
+        }
+      }
+      const observer = new TestObserver();
+      await delay();
+      audioVideoController.addObserver(observer);
+      event += 1;
+      audioVideoController.forEachObserver((observer: AudioVideoObserver) => {
+        observer.encodingSimulcastLayersDidChange(SimulcastLayers.High);
+      });
+      await delay();
+      audioVideoController.removeObserver(observer);
+      event += 1;
+      audioVideoController.forEachObserver((observer: AudioVideoObserver) => {
+        observer.encodingSimulcastLayersDidChange(SimulcastLayers.High);
+      });
+      await delay();
+      audioVideoController.addObserver(observer);
+      event += 1;
+      audioVideoController.forEachObserver((observer: AudioVideoObserver) => {
+        observer.encodingSimulcastLayersDidChange(SimulcastLayers.High);
+      });
+      await delay();
+      event += 1;
+      audioVideoController.forEachObserver((observer: AudioVideoObserver) => {
+        observer.encodingSimulcastLayersDidChange(SimulcastLayers.High);
+      });
+      audioVideoController.removeObserver(observer);
+      await delay();
+      audioVideoController.addObserver(observer);
+      event += 1;
+      audioVideoController.forEachObserver((observer: AudioVideoObserver) => {
+        observer.encodingSimulcastLayersDidChange(SimulcastLayers.High);
+      });
+      await delay();
+      expect(event).to.equal(5);
+      expect(observed).to.equal(3);
+    });
+
+    it('observer should get called', async () => {
+      configuration.enableUnifiedPlanForChromiumBasedBrowsers = true;
+      configuration.enableSimulcastForUnifiedPlanChromiumBasedBrowsers = true;
+      domMockBehavior.browserName = 'chrome';
+      domMockBuilder = new DOMMockBuilder(domMockBehavior);
+      audioVideoController = new DefaultAudioVideoController(
+        configuration,
+        new NoOpDebugLogger(),
+        webSocketAdapter,
+        new NoOpDeviceController(),
+        reconnectController
+      );
+      class TestObserver implements AudioVideoObserver {
+        encodingSimulcastLayersDidChange(_simulcastLayers: SimulcastLayers): void {}
+      }
+      const observer = new TestObserver();
+      const spy = sinon.spy(observer, 'encodingSimulcastLayersDidChange');
+      audioVideoController.addObserver(observer);
+      await start();
+      await delay();
+      // @ts-ignore
+      audioVideoController.meetingSessionContext.videoUplinkBandwidthPolicy.chooseEncodingParameters();
+      await delay();
+      await stop();
+      expect(spy.calledOnce).to.be.true;
+    });
+
+    it('observer should not get called with non-simulcast policy', async () => {
+      audioVideoController = new DefaultAudioVideoController(
+        configuration,
+        new NoOpDebugLogger(),
+        webSocketAdapter,
+        new NoOpDeviceController(),
+        reconnectController
+      );
+      class TestObserver implements AudioVideoObserver {
+        encodingSimulcastLayersDidChange(_simulcastLayers: SimulcastLayers): void {}
+      }
+      const observer = new TestObserver();
+      const spy = sinon.spy(observer, 'encodingSimulcastLayersDidChange');
+      audioVideoController.addObserver(observer);
+      await start();
+      await delay();
+      // @ts-ignore
+      audioVideoController.meetingSessionContext.videoUplinkBandwidthPolicy.chooseEncodingParameters();
+      await delay();
+      await stop();
+      expect(spy.notCalled).to.be.true;
     });
   });
 });