+++ test/videostreamindex/DefaultVideoStreamIndex.test.ts
@@ -14,6 +14,7 @@ import {
   SdkSubscribeAckFrame,
   SdkTrackMapping,
 } from '../../src/signalingprotocol/SignalingProtocol.js';
+import VideoSource from '../../src/videosource/VideoSource';
 import DefaultVideoStreamIndex from '../../src/videostreamindex/DefaultVideoStreamIndex';
 
 describe('DefaultVideoStreamIndex', () => {
@@ -220,11 +221,14 @@ describe('DefaultVideoStreamIndex', () => {
     });
   });
 
-  describe('allVideoSendingAttendeesExcludingSelf', () => {
-    it('returns set of all video sending attendees', () => {
-      expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy1'))).to.deep.equal([]);
+  describe('allVideoSendingSourcesExcludingSelf', () => {
+    const compare = (a: VideoSource, b: VideoSource): number =>
+      a.attendee.attendeeId.localeCompare(b.attendee.attendeeId);
+
+    it('returns set of all video sources', () => {
+      expect(index.allVideoSendingSourcesExcludingSelf('xy1')).to.deep.equal([]);
       index.integrateIndexFrame(new SdkIndexFrame());
-      expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy1'))).to.deep.equal([]);
+      expect(index.allVideoSendingSourcesExcludingSelf('xy1')).to.deep.equal([]);
       index.integrateIndexFrame(
         new SdkIndexFrame({
           sources: [
@@ -233,6 +237,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 1,
               maxBitrateKbps: 100,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -240,6 +245,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 1,
               maxBitrateKbps: 600,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.AUDIO,
             }),
             new SdkStreamDescriptor({
@@ -247,6 +253,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 399,
               maxBitrateKbps: 800,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -254,6 +261,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 2,
               maxBitrateKbps: 50,
               attendeeId: 'xy2',
+              externalUserId: 'ext#xy2',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -261,6 +269,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 399,
               maxBitrateKbps: 200,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -268,16 +277,20 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 2,
               maxBitrateKbps: 400,
               attendeeId: 'xy2',
+              externalUserId: 'ext#xy2',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
           ],
         })
       );
-      expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy1')).sort()).to.deep.equal(
-        ['xy2'].sort()
+      expect(index.allVideoSendingSourcesExcludingSelf('xy1').sort()).to.deep.equal(
+        [{ attendee: { attendeeId: 'xy2', externalUserId: 'ext#xy2' } }].sort(compare)
       );
-      expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy3')).sort()).to.deep.equal(
-        ['xy1', 'xy2'].sort()
+      expect(Array.from(index.allVideoSendingSourcesExcludingSelf('xy3')).sort()).to.deep.equal(
+        [
+          { attendee: { attendeeId: 'xy1', externalUserId: 'ext#xy1' } },
+          { attendee: { attendeeId: 'xy2', externalUserId: 'ext#xy2' } },
+        ].sort(compare)
       );
     });
   });